Enable workflow schedule and add Keepalive job
Github workflows get disabled if there were no commits in the last 60 days (see https://docs.github.com/en/actions/managing-workflow-runs-and-deployments/managing-workflow-runs/disabling-and-enabling-a-workflow).

This can be avoided by enabling the workflow via the API (this resets the timer for the workflow). See https://docs.github.com/en/rest/actions/workflows?apiVersion=2022-11-28#enable-a-workflow for the API endpoint documentation.

This adds a new job to the workflow that just calls this endpoint to make sure the workflow never gets disabled.