Include legacy `addons.json` snapshot
Added a snapshot of the current `addons.json` to the deployed pages. This makes sure that existing applications consuming `addons.json` do not break once we merge all the reworks upstream. I've included a new `_deprecated` property in the `addons.json` to let everyone looking at that file now that they should migrate to the new `manifest.json`.

We should probably update the snapshot once before merging upstream, so it includes all updates to addons between now and then.